pyproject.toml
[project]
name = "tracking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.12.0",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.0",
    "torch>=2.10.0",
    "transformers<4.50",
]
